+++ mullvad-ipc-client/Cargo.toml
@@ -6,3 +6,5 @@ description = "RPC client for Mullvad daemon"
 license = "GPL-3.0"
 
 [dependencies]
+serde = "1.0"
+talpid-ipc = { path = "../talpid-ipc" }
